ESSNTL-2257
With this change, when clicking 'Reset Filters' button in the inventory table, the filter chips go back to the default filters of status fresh and stale, and the URL reflects that.

To see changes navigate to inventory page,
Add a filter (or remove one)
Press the reset filter button
The default filters are restored, and the URL will reflect the parameters

Test had to be updated because now when pressing 'reset all', we are also setting filters.