[ESSNTL-3600] Set name filter as default filter (#1697)
A previous change removed `value` from the operating system filter config for PrimaryToolbar. ConditionalFilter needs this to determine which filter is currently being used. A side-effect was that if one of the filters didn't have the `value` set, it would set it as the default filter.

The previous execution of the code was a bit verbose, and I don't think it was needed. It seems you can hard-code the value in and everything is handled normally.

Co-authored-by: Michael Johnson <micjohns@newbeginnings.usersys.redhat.com>