fix: Disable buttons when not enough permissions
This sets the minimal permissions required in order for users to be able
to press Delete and edit buttons on System details page. If a host is in
a group, then it also checks if user has group level access (which is a
softer reqirement compared to general inventory:hosts:write).