add regex and fix system profile
This commit uses regex to split the versions passed through the url. It also modifies the way we build the system profile structure that is passed via the api to build the request. Now, if there are no RHEL versions, we don't add that object to the system profile, same for CentOS Linux versions.