feat(RHINENG-9248): create bootc tree view (#2178)
* feat(RHINENG-9248): create bootc tree view

To test:
Log into account with systems with bootc data
Go to /insights/inventory url
Click the toggle that is right-aligned in the Systems page header

From this point, you should see a table with expandable rows. Those expanded rows include specific hash data for each image name. The system counts  for the hash commits for the nested table should add up to equal the system counts for the parent image.

Please also ensure that the numbers in the system column line up for the table and nested table.

* feat: Make suggested updates

* feat: Add non bootc system count

Adds count of systems without bootc images to the bootc image table.

---------

Co-authored-by: Michael Johnson <micjohns@micjohns-thinkpadp16vgen1.rmtusnc.csb>