feat(RHINENG-9685): image/package mode icons to inv table
This PR adds Image and Package mode icons to the Inventory table. We use the api to request the `bootc_status` from `system_profile` and if `bootc_status` exists, we show the image mode icon. If it doesn't exist, we show the package mode icon.

Hovering over each will provide a popover with a description of the image and package modes, respectively.