feat(RHINENG-10513): Replace groups with workspaces
This PR implements the feature flag for workspaces. If the feature flag is on, then "Workspace" will show up in place of "Group".

This PR only implements this change in regards to customer-facing use of the word "Group".

To test the flag, either turn the flag on, or open src/Utilities/hooks/useWorkspaceFeatureFlag.js and set the value to true.