fix(RHINENG-12953): strip whitespace edit display and ansible hostname
It is possible to set both Display name and Ansible hostname to a space character, ' '. Not only this makes the UI confusing, it makes you unable to click the host in the Systems page: you have to click Edit, change the name and only then you can open the host details.

Trailing whitespace is cleaned on some places, but not consistently; display name "test   " is stored as such.

To test, find a system and click the system to view its details. Edit the display name and ansible hostname values.

This PR disables the Save button if the name hasn't changed, or is empty, including only whitespaces. Also, if you add whitespace before or after the string, it will trim the whitespace on both sides.