chore(RHINENG-16241): Add help tooltip to RHC status
This PR adds a tooltip to the RHC status on the inventory details
screen. To test, go to Inventory -> System Details and see the RHC
property of the System status card. The property should now have an icon
and the text should match:

"The displayed RHC status indicates that the RHC client is installed and
configured but may not reflect actual connectivity. For further
troubleshooting, please visit Automation Toolkit > Remediations."