fix(RHINENG-16241): Update text in rhc popover
This PR updates popover text to the RHC status on the inventory details screen. To test, go to Inventory -> System Details and see the RHC property of the System status card. The property should now have an icon and the text should match:

"The RHC client was installed and configured but may not reflect actual connectivity.

To view the remediation status of your system, got to Automation Toolkit
> Remediations and open a remediation that your system is associated
with. Under the Systems tab, you will find the Connection Status."