fix(RHINENG-11953): remove 'Update' from workspace details actions (#2527)
Description of Problem
When navigating to Inventory > Workspaces and clicking a workspace to see its details, the dropdown action list include 'Update'. This action is something for systems assigned to the workspace and not an action on the workspace itself. This should be removed from the dropdown.

Steps to Reproduce
- Navigate to Inventory > Workspaces
- Select a Workspace to see its details
- Click on the dropdown for Workspace actions

Co-authored-by: Michael Johnson <micjohns@micjohns-thinkpadp16vgen1.rmtusnc.csb>
Co-authored-by: Zhamilya Abikenova <zabikeno@redhat.com>