fix: remove add system confirmation modal
When on the workspace details page, if you click the "add systems"
button, it will open the modal to add systems to the workspace.

Currently, if you select systems that are already in a workspace, a
warning will populate at the bottom of the modal. Also, it will disable
the "Add systems" button in the modal.

Previously, a user had the ability to bypass this warning and add the
systems anyway via a confirmation modal. But because we disable the
button, there's no need for the confirmation modal. In fact, it's
currently not accessible anyway, so I'm removing them.