fix(RHINENG-19923): properly pass feature flags
The workspace filter on the systems table improperly shows a filter for
"No workspace" instead of the updated "Ungrouped hosts" after Kessel
phase 0.

This is because of a recent PR that was merged that improperly passed
the value of certain flags. With the correct names updated, everything
works as normal.