feat(RHINENG-20062): add SystemsTable action kebab
This PR adds the action kebab to the toolbar of the SystemsTable. To
test select systems in bulk from the table and add to and remove from
workspaces. When a system already belongs to a workspace, only the
Remove from workspace option should be enabled. And when a system
doesn't belong to a workspace, only the Add to workspace option should
be enabled.

Also, test that this functionality still works for the standard
Inventory table, as I moved some of the functionality into a helper. I
tested on my end and everything seemed to work correctly.