chore(RHINENG-20064): integrate row actions SystemsTable
This PR integrates row actions in the new SystemsTable. To test, set
`isSystemsTableEnabled` to true in `src/Routes.js` and load the
Inventory -> Systems page.

The functionality of adding/removing a system from a workspace, editing
the display name of a system, and deleting a system from the inventory
should all work as they did before.

Verify that you can use row actions in conjunction with toolbar actions
and everything acts normally.