fix(RHINENG-20205): delete custom staleness on reset
When a customer resets its custom staleness & deletion to the default values, this change is saved as a 'custom' staleness & deletion record with the same values as 'default'.

This causes unnecessary computation and additional logs that could be avoided by simply removing the 'custom' staleness & deletion for this account and using default.

Steps to Reproduce
- Open the Staleness and Deletion screen where custom values are set
- Click on 'Reset to default setting' and Save

Actual Behavior
The custom staleness & deletion record for the account is updated with default values.

Expected Behavior
The custom staleness & deletion record for the account should be deleted.