chore: bulk update packages
I bulk updated most of the packages from the dependabot package update
PRs.

The biggest change is to a component. I found some issues and decided to
fix them. I also removed a test because it wasn't doing what I think the
writer expected it to. I don't think the rerender function mimics the
idea of opening and closing a modal.