chore: bulk update packages (#2695)
I bulk updated most of the packages from the dependabot package update
PRs.

The biggest change is to a component. I found some issues and decided to
fix them. I also removed a test because it wasn't doing what I think the
writer expected it to. I don't think the rerender function mimics the
idea of opening and closing a modal.

Co-authored-by: Michael Johnson <micjohns@micjohns-thinkpadp16vgen1.rmtusnc.csb>
Co-authored-by: Zhamilya Abikenova <zabikeno@redhat.com>