feat(RHINENG-15792): Add RHEL AI fact to system details
This PR adds a new card for RHEL AI facts on the Details tab of the
System Details page. We're leaving off the tracker for the remaining
memory, per UX.

Check the card for mock-ups of the new RHEL AI facts. Also, notice that
the mock up shows it in 2 columns but that will be implemented in a
later PR.