fix(RHINENG-22300): filter tags modal from systems table
If you:
Navigate to the systems table
Select Tags filter from the filter drop down
Select the x more tags available option at the bottom of the Filter by tags dropdown
try to filter the table by name

It does not filter. This seems to be an issue that has been there for a while.